+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `downloadFcsFiles()` function
 - `downloadAttachment()` function
 - `updatePopulation()` function
+- `applyScaleSet()` function
 - The `authenticate()` function will now prompt for a password if omitted and
   running in RStudio or the `getPass` library is installed.
 - The `authenticate` function now supports two-factor authentication.